Move WT_DATA_HANDLE into its own header file (dhandle.h).
--HG--
rename : src/include/btree.h => src/include/dhandle.h
